Moved TestTestServer to its own file
